# application/api.yaml
openapi: 3.0.1
info:
  description: Api Documentation
  version: '1.0'
  title: Api Documentation
  termsOfService: 'urn:tos'
  contact: {}
paths:
  /exercises:
    get:
      summary: Gets the full list of exercises
      operationId: exercisesIndex
      security:
      - authCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Exercise'
    post:
      summary: Creates a new exercise
      operationId: exercisesAdd
      security:
      - authCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Exercise'
      responses:
        '201':
          description: Exercise has been created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exercise'
        '400':
          description: Bad Request
        '401':
          description: User must be logged in to submit solutions to the exercise
        '403':
          description: User is not allowed to create a new exercise
  '/exercises/{exerciseId}':
    get:
      summary: Gets the specified exercise
      security:
      - authCookie: []
      parameters:
      - name: exerciseId
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exercise'
        '401':
          description: User must be logged in to see this exercise
        '403':
          description: User is not allowed to see this exercise
        '404':
          description: Exercise does not exist
  '/exercises/{exerciseId}/submissions':
    get:
      summary: Gets the list of submissions for the specified exercise
      security:
      - authCookie: []
      parameters:
      - name: exerciseId
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExerciseSubmission'
        '401':
          description: User must be logged in to see the submissions.
        '403':
          description: User is not allowed to see the submissions.
        '404':
          description: Exercise does not exist
    post:
      summary: Submits the code as the current user's solution for the exercise
      security:
      - authCookie: []
      parameters:
      - name: exerciseId
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExerciseSubmission'
      responses:
        '201':
          description: Submission has been created
        '400':
          description: Submitted solution does not fit the exercise parameters
        '401':
          description: User must be logged in to submit solutions to the exercise
        '403':
          description: User is not allowed to post solutions for this exercise
        '404':
          description: Exercise does not exist
  '/exercises/{exerciseId}/test':
    post:
      summary: Validate the code against the exercise's test suite
      security:
      - authCookie: []
      parameters:
      - name: exerciseId
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunnerArguments'
      responses:
        '200':
          description: The test suite has been executed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodeValidationReport'
        '400':
          description: Submitted solution does not fit the exercise parameters
        '401':
          description: User must be logged in to see this exercise
        '403':
          description: User is not allowed to see this exercise
        '404':
          description: Exercise does not exist
  '/exercises/{exerciseId}/submissions/{submissionId}':
    get:
      summary: Gets the list of submissions for the specified exercise
      security:
      - authCookie: []
      parameters:
      - name: exerciseId
        in: path
        required: true
        schema:
          type: integer
      - name: submissionId
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExerciseSubmission'
        '401':
          description: User must be logged in to see this submission
        '403':
          description: User is not allowed to see this submission
        '404':
          description: Submission does not exist
  '/exercises/{exerciseId}/submissions/{submissionId}/results':
    get:
      summary: Gets the results of the exercise validation for this submission
      security:
      - authCookie: []
      parameters:
      - name: exerciseId
        in: path
        required: true
        schema:
          type: integer
      - name: submissionId
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodeValidationReport'
        '401':
          description: User must be logged in to see the results
        '403':
          description: User is not allowed to see the results
        '404':
          description: Submission does not exist
...
servers:
- url: '//localhost:8080/'
components:
  schemas:
    RunnerArguments:
      type: object
      properties:
        code:
          description: code
          type: string
        language:
          description: language
          type: string
      example:
        language: js
        code: 'function main() { console.log(''Hello World''); }'
    CodeValidationReport:
      type: object
      properties:
        exerciseId:
          type: integer
        results:
          type: array
          items:
            type: object
            properties:
              testName:
                type: string
              passed:
                type: boolean
              inputParameters:
                type: array
                items: {}
              expectedOutput: {}
              actualOutput: {}
...
    IntlString:
      type: object
      properties:
        fr:
          type: string
        en:
          type: string
      required:
      - fr
    Exercise:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        title:
          $ref: '#/components/schemas/IntlString'
        description:
          $ref: '#/components/schemas/IntlString'
        difficulty:
          $ref: '#/components/schemas/Difficulties'
        template:
          $ref: '#/components/schemas/Template'
        supportedLanguages:
          type: array
          items:
            type: string
        sampleTestCases:
          type: array
          items:
            $ref: '#/components/schemas/TestCase'
      required:
      - id
      - title
      - description
      - difficulty
      - template
      - sampleTestCases
    ExerciseSubmission:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        exerciseId:
          type: integer
        userId:
          type: integer
        code:
          type: string
        language:
          $ref: '#/components/schemas/'
        #validationReport:
        #  $ref: '#/components/schemas/CodeValidationReport'
        createdAt:
          type: string
          format: date-time
      required:
      - id
      - exerciseId
      - userId
      - code
      - language
    Template:
      type: object
      properties:
        className:
          type: string
        functionName:
          type: string
        params:
          type: array
          items:
            $ref: '#/components/schemas/Parameter'
        functionReturnValue:
          $ref: '#/components/schemas/AnyValue'
        functionReturnType:
          $ref: '#/components/schemas/SupportedType'
        prependedCode:
          type: object
          additionalProperties:
            type: string
        appendedCode:
          type: object
          additionalProperties:
            type: string
      required:
      - functionName
      - functionReturnValue
      - params
      - functionReturnType
      - prependedCode
      - appendedCode
    SupportedType:
      type: string
      enum:
      - BOOLEAN
      - INT
      - FLOAT
      - STRING
      - CHAR
      - BOOLEAN_ARRAY
      - INT_ARRAY
      - FLOAT_ARRAY
      - STRING_ARRAY
      - CHAR_ARRAY
...
    Difficulties:
      description: Exercise difficulty range
      type: string
      enum:
      - EASY
      - MEDIUM
      - HARD
...
    AnyValue:
      description: Type encompassing any value type
      nullable: true
    TestCase:
      description: Test case data structure used to generate examplary test cases
      type: object
      properties:
        it:
          type: string
        assertions:
          type: array
          items:
            $ref: '#/components/schemas/Assertion'
      required:
      - it
      - assertions
    Assertion:
      description: An assertion to be included in a test
      type: object
      properties:
        inputArguments:
          type: array
          items:
            $ref: '#/components/schemas/Argument'
        expectedOutput:
          $ref: '#/components/schemas/Argument'
      required:
      - inputArguments
      - expectedOutput
    Argument:
      description: An argument
      type: object
      properties:
        type:
          $ref: '#/components/schemas/SupportedType'
        value:
          $ref: '#/components/schemas/AnyValue'
      required:
      - type
      - value
    Parameter:
      description: A parameter
      type: object
      properties:
        name:
          type: string
        type:
          $ref: '#/components/schemas/SupportedType'
      required:
      - type
      - name
...
  securitySchemes:
    authCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
...